rename styleContext.js to appContext.js
