+++ client/src/components/admin/config-style.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AppContext from '../../util/styleContext';
+import AppContext from '../../util/appContext';
 import { Row, Col, Button, Alert, Divider, Modal, Form, Input } from 'antd';
 import { ExclamationCircleOutlined } from '@ant-design/icons';
