+++ client/src/components/header.jsx
@@ -3,7 +3,7 @@ import { DownloadOutlined } from '@ant-design/icons';
 import React from 'react';
 import { Link } from 'react-router-dom';
 import SearchBar from './search-bar';
-import AppContext from '../util/styleContext';
+import AppContext from '../util/appContext';
 import AdminButton from './buttons/admin';
 import HelpButton from './buttons/help';
