+++ client/src/pages/search-results.jsx
@@ -4,7 +4,8 @@ import { Layout } from 'antd';
 import DataTable from '../components/data-table';
 import Header from '../components/header';
 import { downloadFileToClient } from '../util';
-import AppContext from '../util/styleContext';
+import AppContext from '../util/appContext';
+
 
 const parseParams = (querystring) => {
   const params = new URLSearchParams(querystring);
